renamed from service to internal (#101)
