+++ microservices/the_monkeys_file_storage/main.go
@@ -6,8 +6,8 @@ import (
     "github.com/the-monkeys/the_monkeys/common"
     "github.com/the-monkeys/the_monkeys/microservices/the_monkeys_file_storage/config"
     "github.com/the-monkeys/the_monkeys/microservices/the_monkeys_file_storage/constant"
-    "github.com/the-monkeys/the_monkeys/microservices/the_monkeys_file_storage/service/pb"
-    "github.com/the-monkeys/the_monkeys/microservices/the_monkeys_file_storage/service/server"
+    "github.com/the-monkeys/the_monkeys/microservices/the_monkeys_file_storage/internal/pb"
+    "github.com/the-monkeys/the_monkeys/microservices/the_monkeys_file_storage/internal/server"
 
     "github.com/sirupsen/logrus"
     "google.golang.org/grpc"
